db load and save method remade and fixed

+++ src/Server/Database/FriendRequestList.cpp
@@ -55,12 +55,12 @@ std::string FriendRequestList::toString(Database &db) const {
 
 void FriendRequestList::write(FILE* file) {
 	size_t size = this->received.size();
-	fwrite(&size, sizeof(int), 1, file);
+	fwrite(&size, sizeof(size_t), 1, file);
 	for (size_t i = 0; i < size; i++) {
 		fwrite(&(this->received[i]), sizeof(int), 1, file);
 	}
 	size = this->sent.size();
-	fwrite(&size, sizeof(int), 1, file);
+	fwrite(&size, sizeof(size_t), 1, file);
 	for (size_t i = 0; i < size; i++) {
 		fwrite(&(this->sent[i]), sizeof(int), 1, file);
 	}
@@ -68,13 +68,13 @@ void FriendRequestList::write(FILE* file) {
 
 void FriendRequestList::read(FILE* file) {
 	size_t size;
-	fread(&size, sizeof(int), 1, file);
+	fread(&size, sizeof(size_t), 1, file);
 	for (size_t i = 0; i < size; i++) {
 		int x;
 		fread(&x, sizeof(int), 1, file);
 		this->received.push_back(x);
 	}
-	fread(&size, sizeof(int), 1, file);
+	fread(&size, sizeof(size_t), 1, file);
 	for (size_t i = 0; i < size; i++) {
 		int x;
 		fread(&x, sizeof(int), 1, file);
